requirements.txt
Django==1.11.3
djangorestframework==3.6.3
psycopg2==2.7.2
pytz==2017.2
requests==2.18.1

